+++ db/001_init.js
@@ -2,8 +2,8 @@ const db = require('./connection'),
 	channels = require('../channels');
 
 db.serialize(function() {
-	db.run("DROP TABLE sandboxes;");
-	db.run("CREATE TABLE IF NOT EXISTS sandboxes(sandbox TEXT PRIMARY KEY ASC, team TEXT, owner TEXT);");
+	db.run("DROP TABLE IF EXISTS sandboxes;");
+	db.run("CREATE TABLE sandboxes(sandbox TEXT PRIMARY KEY ASC, team TEXT, owner TEXT);");
 
 	console.log("Database initialized.");
