+++ package.json
@@ -23,6 +23,7 @@
   "dependencies": {
     "@slack/client": "^3.5.3",
     "bluebird": "^3.3.4",
+    "cron": "^1.3.0",
     "request": "2.76.0",
     "sqlite3": "^3.1.9"
   }
